Fixed bugs

+++ app/src/main/java/com/sgc/speedometer/data/model/SpeedometerRecord.kt
@@ -8,6 +8,6 @@ data class SpeedometerRecord(
     var duration: Date = Date(0),
     var distance: Double = 0.0,
     var averageSpeed: Int = 0,
-    var maxSpeed:Int = 0,
-    var currentSpeed: Int = 0
+    var maxSpeed:Double = 0.0,
+    var currentSpeed: Double = 0.0
 ) : Parcelable
